requirements.txt
--find-links https://download.pytorch.org/whl/cu126
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0

accelerate==0.30.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
colpali_engine==0.3.8
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
GPUtil==1.4.0
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lancedb==0.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mcp==1.8.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
narwhals==1.38.2
networkx==3.4.2
numpy==1.26.4
openai==1.78.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
peft==0.14.0
pillow==10.4.0
propcache==0.3.1
psutil==5.9.8
py==1.11.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pylance==0.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
ratelimiter==1.2.0.post0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retry==0.9.2
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.9
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spaces==0.35.0
sse-starlette==2.3.5
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
tomlkit==0.12.0
tqdm==4.67.1
transformers==4.47.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==11.0.3
yarl==1.20.0
ollama==0.4.8